Add APIs to check username & email availability
