+++ src/resources/resource_users.py
@@ -2,13 +2,15 @@ import uuid as uuid_
 from abc import abstractmethod
 
 from flask_restful import marshal_with, reqparse
-from rubix_http.exceptions.exception import NotFoundException
+from rubix_http.exceptions.exception import NotFoundException, BadDataException
 from rubix_http.resource import RubixResource
+from werkzeug.security import check_password_hash
 
 from src.models.enum import StateType
 from src.models.user.model_user import UserModel
 from src.resources.rest_schema.schema_user import user_all_attributes, user_return_fields, user_all_fields_with_children
-from src.resources.utils import encrypt_password, parse_user_update, encode_jwt_token
+from src.resources.utils import encrypt_password, parse_user_update, encode_jwt_token, get_access_token, \
+    decode_jwt_token
 
 
 class UsersResourceList(RubixResource):
@@ -83,7 +85,6 @@ class UsersResourceByUsername(UsersResource):
 
 
 class UsersVerifyResource(RubixResource):
-
     @classmethod
     def post(cls):
         parser = reqparse.RequestParser()
@@ -95,3 +96,67 @@ class UsersVerifyResource(RubixResource):
         user.state = StateType.VERIFIED
         user.commit()
         return {'message': 'User has been verified successfully'}
+
+
+class UsersChangePasswordResource(RubixResource):
+
+    @classmethod
+    def post(cls):
+        access_token = get_access_token()
+        parser = reqparse.RequestParser()
+        parser.add_argument('new_password', type=str, required=True, store_missing=False)
+        args = parser.parse_args()
+        username = decode_jwt_token(access_token).get('username', '')
+        user: UserModel = UserModel.find_by_username(username)
+        if user is None:
+            raise NotFoundException("User does not exist")
+        user.password = encrypt_password(args['new_password'])
+        user.commit()
+        return {'message': 'Your password has been changed successfully'}
+
+
+class UsersLoginResource(RubixResource):
+    @classmethod
+    def post(cls):
+        parser = reqparse.RequestParser()
+        parser.add_argument('username', type=str, required=True, store_missing=False)
+        parser.add_argument('password', type=str, required=True, store_missing=False)
+        args = parser.parse_args()
+        user = UserModel.find_by_username(args['username'])
+        if user is None:
+            raise NotFoundException('User does not exist')
+        if not check_password_hash(user.password, args['password']):
+            raise BadDataException('username and password combination is incorrect')
+        return encode_jwt_token(user.uuid, user.username)
+
+
+class UsersCheckByUsernameResource(RubixResource):
+    @classmethod
+    def post(cls):
+        parser = reqparse.RequestParser()
+        parser.add_argument('username', type=str, required=True, store_missing=False)
+        args = parser.parse_args()
+        user: UserModel = UserModel.find_by_username(args['username'])
+        return {'exist': bool(user)}
+
+
+class UsersCheckByEmailResource(RubixResource):
+    @classmethod
+    def post(cls):
+        parser = reqparse.RequestParser()
+        parser.add_argument('email', type=str, required=True, store_missing=False)
+        args = parser.parse_args()
+        user: UserModel = UserModel.find_by_email(args['email'])
+        return {'exist': bool(user)}
+
+
+class UsersAuthenticateResource(RubixResource):
+    @classmethod
+    def get(cls):
+        access_token = get_access_token()
+        username = decode_jwt_token(access_token).get('username', '')
+        uuid = decode_jwt_token(access_token).get('sub', '')
+        user: UserModel = UserModel.find_by_uuid(uuid)
+        if user is None:
+            raise NotFoundException("User does not exist")
+        return {'username': username, 'state': user.state.value}
